refactor: simplify project architecture
The project namespace hierarchy should be simplified.  The intent of some of the namespaces and classes was not always clear.

BREAKING CHANGE: All classes in the `Routing` namespace have been moved to the `Http` namespace.  `ModelParamResolver` has been renamed to `ModelResolver`.  The `ModelResolver` now takes the parsed request body as the second constructor argument instead of the server request.  The `PrimaryHandler` has been renamed to `MiddlewareQueueRequestHandler`.  Issue #174.